finish-postcontroller

+++ controllers/postControllers.js
@@ -53,7 +53,7 @@ const createPost = async (req, res, next) => {
     Route: api/posts
     Protected: No
 */
-const getPost = async (req, res, next) => {
+const getPosts = async (req, res, next) => {
     try {
         const posts = await Post.find().sort({ updateAt: -1 })
         res.status(200).json(posts)
@@ -67,7 +67,7 @@ const getPost = async (req, res, next) => {
     Route: api/posts/:id
     Protected: No
 */
-const getPosts = async (req, res, next) => {
+const getPost = async (req, res, next) => {
     try {
         const postId = req.params.id;
         const post = await Post.findById(postId);
@@ -103,7 +103,7 @@ const getCatPosts = async (req, res, next) => {
 const getUserPosts = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const posts = await find({ creator: id }).sort({ createdAt: -1 })
+        const posts = await Post.find({ creator: id }).sort({ createdAt: -1 })
         res.status(200).json(posts);
     } catch (error) {
         return next(new HttpError(error));
@@ -124,37 +124,40 @@ const editPost = async (req, res, next) => {
         let { title, category, description } = req.body;
 
         //ReactQuill has a paragraph opening and closing tag with a break tag in between so there are 11 characters in there already
-        if (!title || category || description.length < 12) {
+        if (!title || !category || description.length < 12) {
             return next(new HttpError('Fill in all fields', 422));
         }
-
-        if (!req.files) {
-            updatedPost = await Post.findByIdAndUpdate(postId, { title, category, description }, { new: true })
-        } else {
-            //get old post from db
-            const oldPost = await Post.findById(postId);
-            fs.unlink(path.join(__dirname, '..', 'uploads', oldPost.thumbnail), async (err) => {
-                if (err) {
-                    return next(new HttpError(err));
+        //get old post from db
+        const oldPost = await Post.findById(postId);
+        if (req.user.id == oldPost.creator.toString()) {
+            if (!req.files) {
+                updatedPost = await Post.findByIdAndUpdate(postId, { title, category, description }, { new: true })
+            } else {
+                //delete old thumbnail from upload
+                fs.unlink(path.join(__dirname, '..', 'uploads', oldPost.thumbnail), async (err) => {
+                    if (err) {
+                        return next(new HttpError(err));
+                    }
+                });
+                //upload new thumbnail
+                const { thumbnail } = req.files;
+                //check file size
+                if (thumbnail.size > 2000000) {
+                    return next(new HttpError('Thumbnail too big. Should be less than 2MB', 422));
                 }
-            });
-            //upload new thumbnail
-            const { thumbnail } = req.files;
-            //check file size
-            if (thumbnail.size > 2000000) {
-                return next(new HttpError('Thumbnail too big. Should be less than 2MB', 422));
+                fileName = thumbnail.name;
+                let splittedFilename = fileName.split('.');
+                newFilename = splittedFilename[0] + uuid() + '.' + splittedFilename[splittedFilename.length - 1];
+                thumbnail.mv(path.join(__dirname, '..', 'uploads', newFilename), async (err) => {
+                    if (err) {
+                        return next(new HttpError(err));
+                    }
+                })
+
+                updatedPost = await Post.findByIdAndUpdate(postId, { title, category, description, thumbnail: newFilename }, { new: true });
             }
-            fileName = thumbnail.name;
-            splittedFilename = fileName.split('.');
-            newFilename = splittedFilename[0] + uuid() + '.' + splittedFilename[splittedFilename.length - 1];
-            thumbnail.mv(path.join(__dirname, '..', 'uploads', newFilename), async (err) => {
-                if (err) {
-                    return next(new HttpError(err));
-                }
-            })
-
-            updatedPost = await Post.findByIdAndUpdate(postId, { title, category, description, thumbnail: newFilename }, { new: true });
         }
+
         if (!updatedPost) {
             return next(new HttpError('Post could not be edited', 422));
         }
@@ -172,26 +175,32 @@ const editPost = async (req, res, next) => {
 */
 const deletePost = async (req, res, next) => {
     try {
-        const postId = req.params;
+        const postId = req.params.id;
         if (!postId) {
             return next(new HttpError('Post unavailable', 404));
         }
 
         const post = await Post.findById(postId);
         const fileName = post?.thumbnail;
-        //delete thumbnail from upload folder
-        fs.unlink(path.join(__dirname, '..', 'uploads', fileName), async (err) => {
-            if (err) {
-                return next(new HttpError(err));
-            } else {
-                await Post.findByIdAndDelete(postId);
-                //find user and reduce post count by 1
-                const currentUser = await User.findById(req.user.id);
-                const userPostCount = currentUser?.posts - 1;
-                await User.findByIdAndUpdate(req.user.id, { posts: userPostCount });
-            }
-        });
-        res.status(200).json({ message: 'Post deleted successfully' });
+        if (req.user.id == post.creator.toString()) {
+            //delete thumbnail from upload folder
+            fs.unlink(path.join(__dirname, '..', 'uploads', fileName), async (err) => {
+                if (err) {
+                    return next(new HttpError(err));
+                } else {
+                    await Post.findByIdAndDelete(postId);
+                    //find user and reduce post count by 1
+                    const currentUser = await User.findById(req.user.id);
+                    const userPostCount = currentUser?.posts - 1;
+                    await User.findByIdAndUpdate(req.user.id, { posts: userPostCount });
+                    res.json({ message: 'Post deleted successfully' });
+                }
+            });
+        } else {
+            return next(new HttpError('Post could not be deleted', 403));
+        }
+
+
     } catch (error) {
         return next(new HttpError(error));
     }
